requirements.txt
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.21.3
google-api-python-client==2.100.0
gspread==5.11.3
google-auth==2.23.3